Add ext to filename when user forget it when rename
This is one solution among two I think.
The second one is change visibility of files the user can choose
from when opening saved workflow.
The first is simpler I guess and it seems more reasonable.

Possibility about change type of `filename` when passed as `pathlib.Path`,
but this operation here is safe. Casting will not raise exception when
`filename` is `pathlib.Path` type indeed.